Working pages
- added working subclasses in category page
- bootstrap5 support
- add poetry
- add darkmode
