--- Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-pelican = "*"
-invoke = "*"
-pyyaml = "*"
-livereload = "*"
-
-[dev-packages]
-
-[requires]
-python_version = "3.10"
