+++ dnd_helper/themes/DnD/templates/index.html
@@ -1,0 +1,7 @@
+{% extends "base.html" %} {% block title %}{{ SITENAME }}{% endblock %} {% block
+content %}
+<div class="row">
+  <div class="col-4 d-flex justify-content-center">
+  </div>
+</div>
+{% endblock %}
