+++ dnd_helper/themes/DnD/templates/spells.html
@@ -1,0 +1,12 @@
+{% extends "base.html" %}
+{% block title %}{{ SITENAME }}{% endblock %}
+{% block content %}
+<div class="container">
+    <div class="col-4 d-flex justify-content-center">
+        <h1>{{ SITENAME }}</h1>
+    </div>
+    {% for cat, null in categories %}
+    <li{% if cat==category %} class="active" {% endif %}><a href="{{ SITEURL }}/{{ cat.url }}">{{ cat }}</a></li>
+        {% endfor %}
+</div>
+{% endblock %}
