+++ dnd_helper/themes/DnD/templates/tag_circle.html
@@ -1,0 +1,13 @@
+<div class="col-md-8 align-self-center">
+    {% if "0" in tag.name %}
+    <h1 class="my-4 text-center">Sztuczki:</h1>
+    {% else %}
+    <h1 class="my-4 text-center">Czary {{ tag.name.split()[1] }} kręgu</h1>
+    {% endif %}
+    <ul>
+        {% for article in articles %}
+        <li class="active"><a href="{{ SITEURL }}/{{ article.url }}">{{ article.title }}</a>
+        </li>
+        {% endfor %}
+    </ul>
+</div>
