+++ dnd_helper/themes/DnD/templates/tag_rarity.html
@@ -1,0 +1,9 @@
+<div class="col-md-8 align-self-center">
+    <h1 class="my-4 text-center">Przedmioty o rzadkości {{ tag.display_name }}:</h1>
+    <ul>
+        {% for article in articles %}
+        <li class="active"><a href="{{ SITEURL }}/{{ article.url }}">{{ article.title }}</a>
+        </li>
+        {% endfor %}
+    </ul>
+</div>
