--- themes/DnD/static/css/style.css
@@ -1,29 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Josefin+Sans&display=swap');
-
-body {
-    font-family: 'Josefin Sans', sans-serif;
-    font-size: 100%;
-    background-color: white;
-    color: #111;
-    padding: 1em;
-    margin: 5% 10%;
-    display: block;
-}
-
-pre {
-    margin: 2em 1em 2em 4em;
-}
-
-#menu li {
-    display: inline;
-}
-
-#post-list {
-    margin-bottom: 1em;
-    margin-top: 1em;
-}
-
-
-.spell {
-    background: #FDF1DC;
-}
