--- themes/DnD/templates/article.html
@@ -1,92 +1,0 @@
-{% extends "base.html" %}
-{% block html_lang %}{{ article.lang }}{% endblock %}
-
-{% block title %}{{ SITENAME }} - {{ article.nazwa|striptags }}{% endblock %}
-
-{% block head %}
-{{ super() }}
-
-{% import 'translations.html' as translations with context %}
-{% if translations.entry_hreflang(article) %}
-{{ translations.entry_hreflang(article) }}
-{% endif %}
-
-{% if article.description %}
-<meta name="description" content="{{ article.description }}" />
-{% endif %}
-
-{% for tag in article.tags %}
-<meta name="tags" content="{{ tag }}" />
-{% endfor %}
-
-{% endblock %}
-
-{% block content %}
-<header>
-  <h1 class="entry-title">
-    <a href="{{ SITEURL }}/{{ article.url }}" rel="bookmark" title="Permalink to {{ article.nazwa|striptags }}">{{
-      article.nazwa }}</a>
-  </h1>
-  {% import 'translations.html' as translations with context %}
-  {{ translations.translations_for(article) }}
-</header>
-<footer class="post-info">
-  {% if article.category %}
-  <div class="category">
-    Category: <a href="{{ SITEURL }}/{{ article.category.url }}">{{ article.category }}</a>
-  </div>
-  {% endif %}
-  {% if article.tags %}
-  <div class="tags">
-    Tags:
-    {% for tag in article.tags %}
-    <a href="{{ SITEURL }}/{{ tag.url }}">{{ tag }}</a>
-    {% endfor %}
-  </div>
-  {% endif %}
-</footer><!-- /.post-info -->
-<article>
-  <div class="col d-flex justify-content-center">
-    <div class="container border rounded spell">
-      <div class="row">
-        <h1 class="spell-name col d-flex justify-content-center">
-          {{ article.nazwa }}
-        </h1>
-      </div>
-      <div class="row">
-        <div class="col">Szkoła: <strong>{{ article["szkoła"] }}</strong></div>
-        <div class="col">Krąg: <strong>{{ article["krąg"] }}</strong></div>
-      </div>
-      <div class="row">
-        <div class="col">
-          Czas rzucania:
-          <strong>{{ article["czas rzucania"] }} {% if article["rytuał"] %} (rytuał) {% endif
-            %}
-          </strong>
-        </div>
-        <div class="col">Zasięg: <strong>{{ article["zasięg"] }}</strong></div>
-      </div>
-      <div class="row">
-        <div class="col">
-          Komponenty: <strong>{{ article["komponenty"] }}</strong>
-        </div>
-        <div class="col">
-          Czas trwania: <strong>{{ article["czas trwania"] }}</strong>
-        </div>
-      </div>
-      <br />
-      <div class="row">
-        <div class="col">{{ article["opis"] }}</div>
-      </div>
-      <div class="row">
-        <div class="col d-flex justify-content-center">
-          {% for klasa in article["klasa"] %}
-          <h2><a href="#" class="badge badge-secondary">{{ klasa }}</a></h2>
-          &nbsp;
-          {% endfor %}
-        </div>
-      </div>
-    </div>
-  </div>
-</article>
-{% endblock %}
