--- themes/DnD/templates/category.html
@@ -1,62 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}{{ SITENAME|striptags }} - {{ category }} category{% endblock %}
-
-{% block content %}
-<div class="row justify-content-center">
-    <div class="col-md-8 align-self-center">
-        <h1 class="mb-1 text-center">{{ category|capitalize }}</h1>
-    </div>
-</div>
-<div class="row justify-content-center">
-    {% if category == "czary" %}
-    <div class="col-md-4 align-self-center">
-        <h2>Klasa:</h2></br>
-        <div class="list-group list-group-flush">
-            <a href="/tag/czar-bard.html" class="list-group-item list-group-item-action">Bard</a>
-            <a href="/tag/czar-kaplan.html" class="list-group-item list-group-item-action">Kapłan</a>
-            <a href="/tag/czar-druid.html" class="list-group-item list-group-item-action">Druid</a>
-            <a href="/tag/czar-wojownik.html" class="list-group-item list-group-item-action">Wojownik</a>
-            <a href="/tag/czar-mnich.html" class="list-group-item list-group-item-action">Mnich</a>
-            <a href="/tag/czar-paladyn.html" class="list-group-item list-group-item-action">Paladyn</a>
-            <a href="/tag/czar-lowca.html" class="list-group-item list-group-item-action">Łowca</a>
-            <a href="/tag/czar-lotr.html" class="list-group-item list-group-item-action">Łotr</a>
-            <a href="/tag/czar-czarownik.html" class="list-group-item list-group-item-action">Czarownik</a>
-            <a href="/tag/czar-mag.html" class="list-group-item list-group-item-action">Mag</a>
-        </div>
-    </div>
-    <div class="col-md-4 align-self-center">
-        <h2>Krąg:</h2></br>
-        <div class="list-group list-group-flush">
-            <a href="/tag/krag-0.html" class="list-group-item list-group-item-action">0 (sztuczki)</a>
-            <a href="/tag/krag-1.html" class="list-group-item list-group-item-action">1</a>
-            <a href="/tag/krag-2.html" class="list-group-item list-group-item-action">2</a>
-            <a href="/tag/krag-3.html" class="list-group-item list-group-item-action">3</a>
-            <a href="/tag/krag-4.html" class="list-group-item list-group-item-action">4</a>
-            <a href="/tag/krag-5.html" class="list-group-item list-group-item-action">5</a>
-            <a href="/tag/krag-6.html" class="list-group-item list-group-item-action">6</a>
-            <a href="/tag/krag-7.html" class="list-group-item list-group-item-action">7</a>
-            <a href="/tag/krag-8.html" class="list-group-item list-group-item-action">8</a>
-            <a href="/tag/krag-9.html" class="list-group-item list-group-item-action">9</a>
-        </div>
-    </div>
-    {% endif %}
-    {% if category == "przedmioty" %}
-    <div class="col-md-8 align-self-center">
-        <h2>Rzadkość:</h2></br>
-        <div class="list-group list-group-flush">
-            <a href="/tag/zwykly.html" class="list-group-item list-group-item-action">Zwykły</a>
-            <a href="/tag/niezwykly.html" class="list-group-item list-group-item-action"
-                style="color: green;">Niezwykły</a>
-            <a href="/tag/rzadki.html" class="list-group-item list-group-item-action" style="color: blue;">Rzadki</a>
-            <a href="/tag/bardzo-rzadki.html" class="list-group-item list-group-item-action"
-                style="color: purple;">Bardzo rzadki</a>
-            <a href="/tag/legendarny.html" class="list-group-item list-group-item-action"
-                style="color: orange;">Legendarny</a>
-            <a href="/tag/artefakt.html" class="list-group-item list-group-item-action"
-                style="color: #dfbe0e;">Artefakt</a>
-        </div>
-    </div>
-    {% endif %}
-</div>
-{% endblock %}
