--- themes/DnD/templates/index.html
@@ -1,7 +1,0 @@
-{% extends "base.html" %} {% block title %}{{ SITENAME }}{% endblock %} {% block
-content %}
-<div class="row">
-  <div class="col-4 d-flex justify-content-center">
-  </div>
-</div>
-{% endblock %}
