--- themes/DnD/templates/item.html
@@ -1,54 +1,0 @@
-{% extends "base.html" %} {% block title %}{{ article.nazwa }} {% endblock %}
-{%block content %}
-<div class="row justify-content-center">
-  <div class="col-md-8 align-self-center d-flex justify-content-center">
-    <div class="container">
-      <div class="row">
-        <h1 class="spell-name col d-flex justify-content-center">
-          {{ article.nazwa }}
-        </h1>
-      </div>
-      <div class="row">
-        <div class="col">Typ: <strong>{{ article["typ"] }}</strong></div>
-        <div class="col">
-          Rzadkość:
-          {% for tag in article.tags %}
-          {% if article["rzadkość"] in tag.display_name %}
-          {% if article["rzadkość"] == "zwykły" %}
-          {% set color = "#495057" %}
-          {% elif article["rzadkość"] == "niezwykły" %}
-          {% set color = "green" %}
-          {% elif article["rzadkość"] == "rzadki" %}
-          {% set color = "blue" %}
-          {% elif article["rzadkość"] == "bardzo rzadki" %}
-          {% set color = "purple" %}
-          {% elif article["rzadkość"] == "legendarny" %}
-          {% set color = "orange" %}
-          {% elif article["rzadkość"] == "artefakt" %}
-          {% set color = "#dfbe0e" %}
-          {% endif %}
-          <a href="{{ tag.url }}" style="color: {{ color }};">
-            <strong>{{ article["rzadkość"] }}</strong>
-          </a>
-          {% endif %}
-          {% endfor %}
-        </div>
-      </div>
-      <div class="row">
-        <div class="col"> Wymaga dostrojenia:
-          {% if article["dostrojenie"] %}
-          <i class="fa-solid fa-check"></i>
-          {% else %}
-          <i class="fa-solid fa-x"></i>
-          {% endif %}
-        </div>
-        <div class="col">Cena: <strong>{{ article["cena"] }}</strong></div>
-      </div>
-      <br />
-      <div class="row">
-        <div class="col">{{ article["opis"] }}</div>
-      </div>
-      {% endblock %}
-    </div>
-  </div>
-</div>
