--- themes/DnD/templates/tag_class.html
@@ -1,19 +1,0 @@
-<h1 class="mb-1 text-center">Czary {{ tag.name.split()[1] }}</h1>
-
-{% for level, articles in articles|groupby(attribute='krąg') %}
-{% if level == 0 %}
-{% set circle = "Sztuczki" %}
-{% else %}
-{% set circle = "Krąg "~level %}
-{% endif %}
-<h3 class="h3 d-inline-block">{{ circle }}</h3>
-<span class="badge badge-dark align-top">{{ articles|count }}</span>
-
-<ul>
-    {% for article in articles %}
-    <li class="active"><a href="{{ SITEURL }}/{{ article.url }}">{{ article.title }}</a>
-    </li>
-    {% endfor %}
-
-</ul>
-{% endfor %}
