--- themes/DnD/templates/tags.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}{{ SITENAME|striptags }} - Tags{% endblock %}
-
-{% block content %}
-<h1>Tags for {{ SITENAME }}</h1>
-<ul>
-    {% for tag, articles in tags %}
-    <li><a href="{{ SITEURL }}/{{ tag.url }}">{{ tag }}</a> ({{ articles|count }})</li>
-    {% endfor %}
-</ul>
-{% endblock %}
